consolidate shared code from various Install.py files into install_helpers.py
