+++ lib/latte/Install.py
@@ -4,7 +4,9 @@
 # used to automate the steps described in the README file in this dir
 
 from __future__ import print_function
-import sys,os,re,subprocess,hashlib,shutil
+import sys,os,re,subprocess,shutil
+sys.path.append('..')
+from install_helpers import error,get_cpus,fullpath,which,geturl,checkmd5sum
 
 # help message
 
@@ -43,75 +45,11 @@ checksums = { \
         '1.2.1' : '85ac414fdada2d04619c8f936344df14', \
         }
 
-# print error message or help
-
-def error(str=None):
-  if not str: print(help)
-  else: print("ERROR",str)
-  sys.exit()
-
-# expand to full path name
-# process leading '~' or relative path
-
-def fullpath(path):
-  return os.path.abspath(os.path.expanduser(path))
-
-def which(program):
-  def is_exe(fpath):
-    return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
-
-  fpath, fname = os.path.split(program)
-  if fpath:
-    if is_exe(program):
-      return program
-  else:
-    for path in os.environ["PATH"].split(os.pathsep):
-      path = path.strip('"')
-      exe_file = os.path.join(path, program)
-      if is_exe(exe_file):
-        return exe_file
-
-  return None
-
-def geturl(url,fname):
-  success = False
-
-  if which('curl') != None:
-    cmd = 'curl -L -o "%s" %s' % (fname,url)
-    try:
-      subprocess.check_output(cmd,stderr=subprocess.STDOUT,shell=True)
-      success = True
-    except subprocess.CalledProcessError as e:
-      print("Calling curl failed with: %s" % e.output.decode('UTF-8'))
-
-  if not success and which('wget') != None:
-    cmd = 'wget -O "%s" %s' % (fname,url)
-    try:
-      subprocess.check_output(cmd,stderr=subprocess.STDOUT,shell=True)
-      success = True
-    except subprocess.CalledProcessError as e:
-      print("Calling wget failed with: %s" % e.output.decode('UTF-8'))
-
-  if not success:
-    error("Failed to download source code with 'curl' or 'wget'")
-  return
-
-def checkmd5sum(md5sum,fname):
-    with open(fname,'rb') as fh:
-        m = hashlib.md5()
-        while True:
-            data = fh.read(81920)
-            if not data:
-                break
-            m.update(data)
-    fh.close()
-    return m.hexdigest() == md5sum
-
 # parse args
 
 args = sys.argv[1:]
 nargs = len(args)
-if nargs == 0: error()
+if nargs == 0: error(help=help)
 
 homepath = "."
 
@@ -123,7 +61,7 @@ linkflag = True
 iarg = 0
 while iarg < nargs:
   if args[iarg] == "-p":
-    if iarg+2 > nargs: error()
+    if iarg+2 > nargs: error(help=help)
     lattedir = fullpath(args[iarg+1])
     pathflag = True
     iarg += 2
@@ -131,15 +69,15 @@ while iarg < nargs:
     buildflag = True
     iarg += 1
   elif args[iarg] == "-m":
-    if iarg+2 > nargs: error()
+    if iarg+2 > nargs: error(help=help)
     suffix = args[iarg+1]
     suffixflag = True
     iarg += 2
   elif args[iarg] == "-v":
-    if iarg+2 > nargs: error()
+    if iarg+2 > nargs: error(help=help)
     version = args[iarg+1]
     iarg += 2
-  else: error()
+  else: error(help=help)
 
 homedir = "LATTE-%s" % version
 
@@ -177,8 +115,12 @@ if buildflag:
 if buildflag:
   print("Building LATTE ...")
   cmd = 'cd "%s"; make' % lattedir
-  txt = subprocess.check_output(cmd,stderr=subprocess.STDOUT,shell=True)
-  print(txt.decode('UTF-8'))
+  try:
+    txt = subprocess.check_output(cmd,stderr=subprocess.STDOUT,shell=True)
+    print(txt.decode('UTF-8'))
+  except subprocess.CalledProcessError as e:
+    print("Make failed with:\n %s" % e.output.decode('UTF-8'))
+    sys.exit(1)
 
 # create 3 links in lib/latte to LATTE dirs
 # do this -b or -p is set
